requirements.txt
django==4.1.3
django-crispy-forms==1.14
django-allauth==0.51